Move zap sender extraction into NIP-57 protocol
