+++ web/src/lib/EventHelper.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { aTagContent, getZapperPubkey, isLegacyEncryption, parseAddress } from './EventHelper';
+import { aTagContent, isLegacyEncryption, parseAddress } from './EventHelper';
 import { generateSecretKey, getPublicKey, nip04, nip44 } from 'nostr-tools';
 
 describe('aTagContent', () => {
@@ -11,65 +11,6 @@ describe('aTagContent', () => {
 		expect(aTagContent({ ...base, kind: 30001, tags: [['d', 'bookmark']] })).toBe(
 			'30001:pubkey:bookmark'
 		);
-	});
-});
-
-describe('getZapperPubkey', () => {
-	it('P tag', () => {
-		expect(
-			getZapperPubkey({
-				pubkey: 'wallet',
-				kind: 9735,
-				content: '',
-				tags: [
-					['p', 'author'],
-					['P', 'zapper']
-				],
-				created_at: 0,
-				id: '',
-				sig: ''
-			})
-		).toBe('zapper');
-	});
-	it('description pubkey', () => {
-		expect(
-			getZapperPubkey({
-				pubkey: 'wallet',
-				kind: 9735,
-				content: '',
-				tags: [
-					['p', 'author'],
-					['description', JSON.stringify({ pubkey: 'zapper' })]
-				],
-				created_at: 0,
-				id: '',
-				sig: ''
-			})
-		).toBe('zapper');
-	});
-	it('none', () => {
-		expect(
-			getZapperPubkey({
-				pubkey: 'wallet',
-				kind: 9735,
-				content: '',
-				tags: [['p', 'author']],
-				created_at: 0,
-				id: '',
-				sig: ''
-			})
-		).toBe(undefined);
-		expect(
-			getZapperPubkey({
-				pubkey: 'wallet',
-				kind: 1,
-				content: '',
-				tags: [['p', 'author']],
-				created_at: 0,
-				id: '',
-				sig: ''
-			})
-		).toBe(undefined);
 	});
 });
