+++ web/src/lib/EventHelper.ts
@@ -103,30 +103,6 @@ export function aTagContent(event: Event): string {
 	return `${event.kind}:${event.pubkey}:${identifier}`;
 }
 
-export function getZapperPubkey(event: Event): string | undefined {
-	if (event.kind !== 9735) {
-		return undefined;
-	}
-
-	const pubkey = filterTags('P', event.tags).at(0);
-	if (pubkey !== undefined) {
-		return pubkey;
-	}
-
-	const description = filterTags('description', event.tags).at(0);
-	if (description === undefined) {
-		return undefined;
-	}
-
-	try {
-		const event9734 = JSON.parse(description) as Event;
-		return event9734.pubkey;
-	} catch (error) {
-		console.warn('[kind 9735 description decode error]', event, error);
-		return undefined;
-	}
-}
-
 export function isNostrHex(hex: string): boolean {
 	return /[0-9a-f]{64}/.test(hex);
 }
